+++ packages/runtime/src/templates/edge/runtime.ts
@@ -45,10 +45,11 @@ declare global {
 globalThis.NFRequestContextMap ||= new Map()
 
 const handler = async (req: Request, context: Context) => {
-  const url = new URL(req.url)
-  if (url.pathname.startsWith('/_next/static/')) {
+  if (Deno.env.get('NETLIFY_DEV')) {
+    // Don't run in dev
     return
   }
+  const url = new URL(req.url)
 
   const geo = {
     country: context.geo.country?.code,