+++ test/index.js
@@ -1,6 +1,6 @@
-jest.mock('../plugin/src/helpers/utils', () => {
+jest.mock('../packages/runtime/src/helpers/utils', () => {
   return {
-    ...jest.requireActual('../plugin/src/helpers/utils'),
+    ...jest.requireActual('../packages/runtime/src/helpers/utils'),
     isNextAuthInstalled: jest.fn(),
   }
 })
@@ -12,11 +12,12 @@ const process = require('process')
 const os = require('os')
 const cpy = require('cpy')
 const { dir: getTmpDir } = require('tmp-promise')
-const { downloadFile } = require('../plugin/src/templates/handlerUtils')
-const { getApiRouteConfigs } = require('../plugin/src/helpers/functions')
-const plugin = require('../plugin/src')
+const { downloadFile } = require('../packages/runtime/src/templates/handlerUtils')
+const { getApiRouteConfigs } = require('../packages/runtime/src/helpers/functions')
+const nextRuntimeFactory = require('../packages/runtime/src')
+const nextRuntime = nextRuntimeFactory({})
 
-const { HANDLER_FUNCTION_NAME, ODB_FUNCTION_NAME, IMAGE_FUNCTION_NAME } = require('../plugin/src/constants')
+const { HANDLER_FUNCTION_NAME, ODB_FUNCTION_NAME, IMAGE_FUNCTION_NAME } = require('../packages/runtime/src/constants')
 const { join } = require('pathe')
 const {
   matchMiddleware,
@@ -25,14 +26,14 @@ const {
   matchesRewrite,
   patchNextFiles,
   unpatchNextFiles,
-} = require('../plugin/src/helpers/files')
+} = require('../packages/runtime/src/helpers/files')
 const {
   getRequiredServerFiles,
   updateRequiredServerFiles,
   generateCustomHeaders,
-} = require('../plugin/src/helpers/config')
+} = require('../packages/runtime/src/helpers/config')
 const { dirname } = require('path')
-const { getProblematicUserRewrites } = require('../plugin/src/helpers/verification')
+const { getProblematicUserRewrites } = require('../packages/runtime/src/helpers/verification')
 
 const chance = new Chance()
 const FIXTURES_DIR = `${__dirname}/fixtures`
@@ -176,23 +177,23 @@ afterEach(async () => {
 describe('preBuild()', () => {
   test('fails if publishing the root of the project', () => {
     defaultArgs.netlifyConfig.build.publish = path.resolve('.')
-    expect(plugin.onPreBuild(defaultArgs)).rejects.toThrowError(
+    expect(nextRuntime.onPreBuild(defaultArgs)).rejects.toThrowError(
       /Your publish directory is pointing to the base directory of your site/,
     )
   })
 
   test('fails if the build version is too old', () => {
     expect(
-      plugin.onPreBuild({
+      nextRuntime.onPreBuild({
         ...defaultArgs,
         constants: { IS_LOCAL: true, NETLIFY_BUILD_VERSION: '18.15.0' },
       }),
-    ).rejects.toThrow('This version of the Essential Next.js plugin requires netlify-cli')
+    ).rejects.toThrow('This version of the Next Runtime requires netlify-cli')
   })
 
   test('passes if the build version is new enough', async () => {
     expect(
-      plugin.onPreBuild({
+      nextRuntime.onPreBuild({
         ...defaultArgs,
         constants: { IS_LOCAL: true, NETLIFY_BUILD_VERSION: '18.16.1' },
       }),
@@ -204,7 +205,7 @@ describe('preBuild()', () => {
 
     const restore = jest.fn()
 
-    await plugin.onPreBuild({
+    await nextRuntime.onPreBuild({
       ...defaultArgs,
       utils: { ...utils, cache: { restore } },
     })
@@ -215,13 +216,13 @@ describe('preBuild()', () => {
   it('forces the target to "server"', async () => {
     const netlifyConfig = { ...defaultArgs.netlifyConfig }
 
-    await plugin.onPreBuild({ ...defaultArgs, netlifyConfig })
+    await nextRuntime.onPreBuild({ ...defaultArgs, netlifyConfig })
     expect(netlifyConfig.build.environment.NEXT_PRIVATE_TARGET).toBe('server')
   })
 })
 
 describe('onBuild()', () => {
-  const { isNextAuthInstalled } = require('../plugin/src/helpers/utils')
+  const { isNextAuthInstalled } = require('../packages/runtime/src/helpers/utils')
 
   beforeEach(() => {
     isNextAuthInstalled.mockImplementation(() => {
@@ -244,7 +245,7 @@ describe('onBuild()', () => {
     initialConfig.config.env.NEXTAUTH_URL = mockUserDefinedSiteUrl
     await updateRequiredServerFiles(netlifyConfig.build.publish, initialConfig)
 
-    await plugin.onBuild(defaultArgs)
+    await nextRuntime.onBuild(defaultArgs)
 
     expect(onBuildHasRun(netlifyConfig)).toBe(true)
     const config = await getRequiredServerFiles(netlifyConfig.build.publish)
@@ -262,7 +263,7 @@ describe('onBuild()', () => {
 
     await moveNextDist()
 
-    await plugin.onBuild(defaultArgs)
+    await nextRuntime.onBuild(defaultArgs)
 
     expect(onBuildHasRun(netlifyConfig)).toBe(true)
     const config = await getRequiredServerFiles(netlifyConfig.build.publish)
@@ -280,7 +281,7 @@ describe('onBuild()', () => {
     initialConfig.config.basePath = '/foo'
     await updateRequiredServerFiles(netlifyConfig.build.publish, initialConfig)
 
-    await plugin.onBuild(defaultArgs)
+    await nextRuntime.onBuild(defaultArgs)
 
     expect(onBuildHasRun(netlifyConfig)).toBe(true)
     const config = await getRequiredServerFiles(netlifyConfig.build.publish)
@@ -294,7 +295,7 @@ describe('onBuild()', () => {
     })
 
     await moveNextDist()
-    await plugin.onBuild(defaultArgs)
+    await nextRuntime.onBuild(defaultArgs)
 
     expect(onBuildHasRun(netlifyConfig)).toBe(true)
     const config = await getRequiredServerFiles(netlifyConfig.build.publish)
@@ -305,7 +306,7 @@ describe('onBuild()', () => {
   test('runs onBuild', async () => {
     await moveNextDist()
 
-    await plugin.onBuild(defaultArgs)
+    await nextRuntime.onBuild(defaultArgs)
 
     expect(onBuildHasRun(netlifyConfig)).toBe(true)
   })
@@ -313,7 +314,7 @@ describe('onBuild()', () => {
   test('skips if NETLIFY_NEXT_PLUGIN_SKIP is set', async () => {
     process.env.NETLIFY_NEXT_PLUGIN_SKIP = 'true'
     await moveNextDist()
-    await plugin.onBuild(defaultArgs)
+    await nextRuntime.onBuild(defaultArgs)
 
     expect(onBuildHasRun(netlifyConfig)).toBe(false)
     delete process.env.NETLIFY_NEXT_PLUGIN_SKIP
@@ -322,7 +323,7 @@ describe('onBuild()', () => {
   test('skips if NEXT_PLUGIN_FORCE_RUN is "false"', async () => {
     process.env.NEXT_PLUGIN_FORCE_RUN = 'false'
     await moveNextDist()
-    await plugin.onBuild(defaultArgs)
+    await nextRuntime.onBuild(defaultArgs)
 
     expect(onBuildHasRun(netlifyConfig)).toBe(false)
     delete process.env.NEXT_PLUGIN_FORCE_RUN
@@ -334,7 +335,7 @@ describe('onBuild()', () => {
     const failBuild = jest.fn().mockImplementation((err) => {
       throw new Error(err)
     })
-    expect(() => plugin.onBuild({ ...defaultArgs, utils: { ...utils, build: { failBuild } } })).rejects.toThrow(
+    expect(() => nextRuntime.onBuild({ ...defaultArgs, utils: { ...utils, build: { failBuild } } })).rejects.toThrow(
       `In most cases it should be set to ".next", unless you have chosen a custom "distDir" in your Next config.`,
     )
     expect(failBuild).toHaveBeenCalled()
@@ -348,7 +349,7 @@ describe('onBuild()', () => {
     })
     netlifyConfig.build.publish = path.resolve('out')
 
-    expect(() => plugin.onBuild({ ...defaultArgs, utils: { ...utils, build: { failBuild } } })).rejects.toThrow(
+    expect(() => nextRuntime.onBuild({ ...defaultArgs, utils: { ...utils, build: { failBuild } } })).rejects.toThrow(
       `Your publish directory is set to "out", but in most cases it should be ".next".`,
     )
     expect(failBuild).toHaveBeenCalled()
@@ -358,14 +359,14 @@ describe('onBuild()', () => {
     await moveNextDist()
     await writeJSON(path.join(process.cwd(), '.next/export-detail.json'), {})
     const failBuild = jest.fn()
-    await plugin.onBuild({ ...defaultArgs, utils: { ...utils, build: { failBuild } } })
+    await nextRuntime.onBuild({ ...defaultArgs, utils: { ...utils, build: { failBuild } } })
     expect(failBuild).toHaveBeenCalled()
   })
 
   test('copy handlers to the internal functions directory', async () => {
     await moveNextDist()
 
-    await plugin.onBuild(defaultArgs)
+    await nextRuntime.onBuild(defaultArgs)
 
     expect(existsSync(`.netlify/functions-internal/___netlify-handler/___netlify-handler.js`)).toBeTruthy()
     expect(existsSync(`.netlify/functions-internal/___netlify-handler/bridge.js`)).toBeTruthy()
@@ -378,7 +379,7 @@ describe('onBuild()', () => {
   test('writes correct redirects to netlifyConfig', async () => {
     await moveNextDist()
 
-    await plugin.onBuild(defaultArgs)
+    await nextRuntime.onBuild(defaultArgs)
     // Not ideal, because it doesn't test precedence, but unfortunately the exact order seems to
     // be non-deterministic, as it depends on filesystem globbing across platforms.
     const sorted = [...netlifyConfig.redirects].sort((a, b) => a.from.localeCompare(b.from))
@@ -388,13 +389,13 @@ describe('onBuild()', () => {
   test('publish dir is/has next dist', async () => {
     await moveNextDist()
 
-    await plugin.onBuild(defaultArgs)
+    await nextRuntime.onBuild(defaultArgs)
     expect(existsSync(path.resolve('.next/BUILD_ID'))).toBeTruthy()
   })
 
   test('generates static files manifest', async () => {
     await moveNextDist()
-    await plugin.onBuild(defaultArgs)
+    await nextRuntime.onBuild(defaultArgs)
     const manifestPath = path.resolve('.next/static-manifest.json')
     expect(existsSync(manifestPath)).toBeTruthy()
     const data = (await readJson(manifestPath)).sort()
@@ -403,7 +404,7 @@ describe('onBuild()', () => {
 
   test('moves static files to root', async () => {
     await moveNextDist()
-    await plugin.onBuild(defaultArgs)
+    await nextRuntime.onBuild(defaultArgs)
     const data = JSON.parse(readFileSync(path.resolve('.next/static-manifest.json'), 'utf8'))
     data.forEach(([_, file]) => {
       expect(existsSync(path.resolve(path.join('.next', file)))).toBeTruthy()
@@ -413,7 +414,7 @@ describe('onBuild()', () => {
 
   test('copies default locale files to top level', async () => {
     await moveNextDist()
-    await plugin.onBuild(defaultArgs)
+    await nextRuntime.onBuild(defaultArgs)
     const data = JSON.parse(readFileSync(path.resolve('.next/static-manifest.json'), 'utf8'))
 
     const locale = 'en/'
@@ -430,7 +431,7 @@ describe('onBuild()', () => {
   // TODO - TO BE MOVED TO TEST AGAINST A PROJECT WITH MIDDLEWARE IN ANOTHER PR
   test.skip('skips static files that match middleware', async () => {
     await moveNextDist()
-    await plugin.onBuild(defaultArgs)
+    await nextRuntime.onBuild(defaultArgs)
 
     expect(existsSync(path.resolve(path.join('.next', 'en', 'middle.html')))).toBeFalsy()
     expect(existsSync(path.resolve(path.join('.next', 'server', 'pages', 'en', 'middle.html')))).toBeTruthy()
@@ -439,7 +440,7 @@ describe('onBuild()', () => {
   test('sets correct config', async () => {
     await moveNextDist()
 
-    await plugin.onBuild(defaultArgs)
+    await nextRuntime.onBuild(defaultArgs)
     const includes = [
       '.env',
       '.env.local',
@@ -473,7 +474,7 @@ describe('onBuild()', () => {
 
   test('generates a file referencing all page sources', async () => {
     await moveNextDist()
-    await plugin.onBuild(defaultArgs)
+    await nextRuntime.onBuild(defaultArgs)
     const handlerPagesFile = path.join(constants.INTERNAL_FUNCTIONS_SRC, HANDLER_FUNCTION_NAME, 'pages.js')
     const odbHandlerPagesFile = path.join(constants.INTERNAL_FUNCTIONS_SRC, ODB_FUNCTION_NAME, 'pages.js')
     expect(existsSync(handlerPagesFile)).toBeTruthy()
@@ -492,7 +493,7 @@ describe('onBuild()', () => {
       netlifyConfig,
       constants: { ...constants, PUBLISH_DIR: dir },
     }
-    await plugin.onBuild(config)
+    await nextRuntime.onBuild(config)
     const handlerPagesFile = path.join(constants.INTERNAL_FUNCTIONS_SRC, HANDLER_FUNCTION_NAME, 'pages.js')
     const odbHandlerPagesFile = path.join(constants.INTERNAL_FUNCTIONS_SRC, ODB_FUNCTION_NAME, 'pages.js')
 
@@ -502,7 +503,7 @@ describe('onBuild()', () => {
 
   test('generates entrypoints with correct references', async () => {
     await moveNextDist()
-    await plugin.onBuild(defaultArgs)
+    await nextRuntime.onBuild(defaultArgs)
 
     const handlerFile = path.join(
       constants.INTERNAL_FUNCTIONS_SRC,
@@ -526,16 +527,37 @@ describe('onBuild()', () => {
     delete routesManifest.staticRoutes
     await writeJSON(manifestPath, routesManifest)
     // The function is supposed to return undefined, but we want to check if it throws
-    expect(await plugin.onBuild(defaultArgs)).toBeUndefined()
+    expect(await nextRuntime.onBuild(defaultArgs)).toBeUndefined()
   })
 
-  test('generates imageconfig file with entries for domains and remotePatterns', async () => {
+  test('generates imageconfig file with entries for domains, remotePatterns, and custom response headers', async () => {
     await moveNextDist()
-    await plugin.onBuild(defaultArgs)
+    const mockHeaderValue = chance.string()
+
+    const updatedArgs = {
+      ...defaultArgs,
+      netlifyConfig: {
+        ...defaultArgs.netlifyConfig,
+        headers: [
+          {
+            for: '/_next/image/',
+            values: {
+              'X-Foo': mockHeaderValue,
+            },
+          },
+        ],
+      },
+    }
+    await nextRuntime.onBuild(updatedArgs)
+
     const imageConfigPath = path.join(constants.INTERNAL_FUNCTIONS_SRC, IMAGE_FUNCTION_NAME, 'imageconfig.json')
     const imageConfigJson = await readJson(imageConfigPath)
+
     expect(imageConfigJson.domains.length).toBe(1)
     expect(imageConfigJson.remotePatterns.length).toBe(1)
+    expect(imageConfigJson.responseHeaders).toStrictEqual({
+      'X-Foo': mockHeaderValue,
+    })
   })
 })
 
@@ -545,7 +567,7 @@ describe('onPostBuild', () => {
 
     const save = jest.fn()
 
-    await plugin.onPostBuild({
+    await nextRuntime.onPostBuild({
       ...defaultArgs,
       utils: { ...utils, cache: { save }, functions: { list: jest.fn().mockResolvedValue([]) } },
     })
@@ -574,7 +596,7 @@ describe('onPostBuild', () => {
     const oldLog = console.log
     const logMock = jest.fn()
     console.log = logMock
-    await plugin.onPostBuild({
+    await nextRuntime.onPostBuild({
       ...defaultArgs,
 
       utils: { ...utils, cache: { save: jest.fn() }, functions: { list } },
@@ -595,10 +617,10 @@ describe('onPostBuild', () => {
     process.env.NETLIFY_NEXT_PLUGIN_SKIP = 'true'
     await moveNextDist()
     const show = jest.fn()
-    await plugin.onPostBuild({ ...defaultArgs, utils: { ...defaultArgs.utils, status: { show } } })
+    await nextRuntime.onPostBuild({ ...defaultArgs, utils: { ...defaultArgs.utils, status: { show } } })
     expect(show).toHaveBeenCalledWith({
       summary: 'Next cache was stored, but all other functions were skipped because NETLIFY_NEXT_PLUGIN_SKIP is set',
-      title: 'Essential Next.js plugin did not run',
+      title: 'Next Runtime did not run',
     })
     delete process.env.NETLIFY_NEXT_PLUGIN_SKIP
   })
@@ -609,11 +631,11 @@ describe('onPostBuild', () => {
     process.env.NEXT_PLUGIN_FORCE_RUN = 'false'
     await moveNextDist()
     const show = jest.fn()
-    await plugin.onPostBuild({ ...defaultArgs, utils: { ...defaultArgs.utils, status: { show } } })
+    await nextRuntime.onPostBuild({ ...defaultArgs, utils: { ...defaultArgs.utils, status: { show } } })
     expect(show).toHaveBeenCalledWith({
       summary:
         'Next cache was stored, but all other functions were skipped because NEXT_PLUGIN_FORCE_RUN is set to false',
-      title: 'Essential Next.js plugin did not run',
+      title: 'Next Runtime did not run',
     })
     delete process.env.NEXT_PLUGIN_FORCE_RUN
   })
@@ -650,7 +672,7 @@ describe('onPostBuild', () => {
 
     const show = jest.fn()
 
-    await plugin.onPostBuild({
+    await nextRuntime.onPostBuild({
       ...defaultArgs,
 
       utils: { ...defaultArgs.utils, status: { show }, functions: { list: jest.fn().mockResolvedValue([]) } },
@@ -746,7 +768,7 @@ describe('onPostBuild', () => {
 
     const show = jest.fn()
 
-    await plugin.onPostBuild({
+    await nextRuntime.onPostBuild({
       ...defaultArgs,
 
       utils: { ...defaultArgs.utils, status: { show }, functions: { list: jest.fn().mockResolvedValue([]) } },
@@ -853,7 +875,7 @@ describe('onPostBuild', () => {
     delete routesManifest.headers
     await writeJSON(manifestPath, routesManifest)
 
-    await plugin.onPostBuild({
+    await nextRuntime.onPostBuild({
       ...defaultArgs,
 
       utils: { ...defaultArgs.utils, status: { show }, functions: { list: jest.fn().mockResolvedValue([]) } },
@@ -1021,7 +1043,7 @@ describe('utility functions', () => {
 describe('function helpers', () => {
   it('downloadFile can download a file', async () => {
     const url =
-      'https://raw.githubusercontent.com/netlify/netlify-plugin-nextjs/c2668af24a78eb69b33222913f44c1900a3bce23/manifest.yml'
+      'https://raw.githubusercontent.com/netlify/next-runtime/c2668af24a78eb69b33222913f44c1900a3bce23/manifest.yml'
     const tmpFile = join(os.tmpdir(), 'next-test', 'downloadfile.txt')
     await ensureDir(path.dirname(tmpFile))
     await downloadFile(url, tmpFile)